pyproject.toml
[project]
name = "temporal-terraform"
version = "0.1.0"
description = "A practical example of using Terraform to configure Temporal Cloud."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-secret-manager==2.23.3",
    "pyyaml==6.0.2",
    "temporalio==1.11.1",
]
